# 8284579: Improve VarHandle checks for interpreter