--- a/src/java.base/share/classes/java/lang/invoke/VarHandle.java
+++ b/src/java.base/share/classes/java/lang/invoke/VarHandle.java
@@ -2064,8 +2064,9 @@ public abstract sealed class VarHandle implements Constable
 
     /**
      * Validates that the given access descriptors method type matches up with
-     * the access mode of this VarHandle. This checking and accessor was merged
-     * to slightly improve efficiency during warmup.
+     * the access mode of this VarHandle, then returns if this is a direct
+     * method handle. These operations were grouped together to slightly
+     * improve efficiency during startup/warmup.
      *
      * @return true if this is a direct VarHandle, false if it's an indirect
      *         VarHandle.