--- a/src/java.base/share/classes/java/lang/invoke/VarHandle.java
+++ b/src/java.base/share/classes/java/lang/invoke/VarHandle.java
@@ -2070,7 +2070,7 @@ public abstract sealed class VarHandle implements Constable
      *
      * @return true if this is a direct VarHandle, false if it's an indirect
      *         VarHandle.
-     * @throws WrongMethodTypeException if there's a access type mismatch
+     * @throws WrongMethodTypeException if there's an access type mismatch
      */
     @ForceInline
     boolean checkAccessModeThenIsDirect(VarHandle.AccessDescriptor ad) {