… src/java.base/share/classes/java/lang/invoke/VarHandleGuards.java