A few additional enhancements aiming to improve VH performance in the interpreter:

- Flatten `TypeAndInvokers`: adds a pointer to `VarHandle` (a small increase 40->48) but removes an object and an indirection on any instance actually used - and might avoid allocating the `MethodHandle[]` unnecessarily on some instances
- Have `checkExactAccessMode` return the directness of the `VarHandle` so that we can avoid some `isDirect` method calls.

Baseline, `-Xint`
```
Benchmark                                 Mode  Cnt    Score   Error  Units
VarHandleExact.exact_exactInvocation      avgt   30  478.324 ? 5.762  ns/op
VarHandleExact.generic_exactInvocation    avgt   30  392.114 ? 1.644  ns/op
VarHandleExact.generic_genericInvocation  avgt   30  822.484 ? 1.865  ns/op
```

Patched, `-Xint`
```
Benchmark                                 Mode  Cnt    Score   Error  Units
VarHandleExact.exact_exactInvocation      avgt   30  437.704 ? 5.320  ns/op
VarHandleExact.generic_exactInvocation    avgt   30  374.512 ? 3.154  ns/op
VarHandleExact.generic_genericInvocation  avgt   30  757.054 ? 1.237  ns/op
```

No significant performance difference in normal mode.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8284579](https://bugs.openjdk.java.net/browse/JDK-8284579): Improve VarHandle checks for interpreter


### Reviewers
 * [Maurizio Cimadamore](https://openjdk.java.net/census#mcimadamore) (@mcimadamore - **Reviewer**) ⚠️ Review applies to 68f414a14a85dc91ded4863c288a42f8be403cbf
 * [Mandy Chung](https://openjdk.java.net/census#mchung) (@mlchung - **Reviewer**) ⚠️ Review applies to 2a4fbd6df78e5b0f2d463f2602016308f641714d


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/8160/head:pull/8160` \
`$ git checkout pull/8160`

Update a local copy of the PR: \
`$ git checkout pull/8160` \
`$ git pull https://git.openjdk.java.net/jdk pull/8160/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 8160`

View PR using the GUI difftool: \
`$ git pr show -t 8160`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/8160.diff">https://git.openjdk.java.net/jdk/pull/8160.diff</a>

</details>