@cl4es 2022-04-08
> TestAdaptVarHandles

Thanks for reviewing, Maurizio. I ran the micros you suggested and I detect no change whatsoever at peak performance. With `-Xint` there are some small improvements in line with the results above.