@dholmes-ora 2022-04-11
`checkExactAccessModeThenIsDirect`?

At a minimum the return value needs documenting via a comment!