@dholmes-ora 2022-04-12
> checkExactAccessMode -> checkAccessModeThenIsDirect

Don't you still want "Exact" in there? That "access" check seems odd anyway as it only checks for one form of mismatch - should it not also check for `!exact && accessModeType(ad.type) == ad.symbolicMethodTypeExact`?